Create type context query
